refactor: float and fixed both use a common macro for nearest_n
