+++ src/common/mod.rs
@@ -1,1 +1,2 @@
+pub(crate) mod generate_nearest_n;
 pub(crate) mod generate_nearest_one;
