+++ src/fixed/query/nearest_n.rs
@@ -1,149 +1,44 @@
-use crate::fixed::kdtree::{Axis, KdTree};
-use crate::fixed::neighbour::Neighbour;
-use crate::types::{Content, Index};
 use az::{Az, Cast};
 use std::collections::BinaryHeap;
 use std::ops::Rem;
+
+use crate::fixed::kdtree::{Axis, KdTree};
+use crate::fixed::neighbour::Neighbour;
+use crate::types::{Content, Index, is_stem_index};
+
+use crate::generate_nearest_n;
 
 impl<A: Axis, T: Content, const K: usize, const B: usize, IDX: Index<T = IDX>>
     KdTree<A, T, K, B, IDX>
 where
     usize: Cast<IDX>,
 {
-    /// Finds the nearest `qty` elements to `query`, using the specified
-    /// distance metric function.
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// use fixed::FixedU16;
-    /// use fixed::types::extra::U0;
-    /// use kiddo::fixed::kdtree::KdTree;
-    /// use kiddo::fixed::distance::squared_euclidean;
-    ///
-    /// type FXD = FixedU16<U0>;
-    ///
-    /// let mut tree: KdTree<FXD, u32, 3, 32, u32> = KdTree::new();
-    ///
-    /// tree.add(&[FXD::from_num(1), FXD::from_num(2), FXD::from_num(5)], 100);
-    /// tree.add(&[FXD::from_num(2), FXD::from_num(3), FXD::from_num(6)], 101);
-    ///
-    /// let nearest: Vec<_> = tree.nearest_n(&[FXD::from_num(1), FXD::from_num(2), FXD::from_num(5)], 1, &squared_euclidean);
-    ///
-    /// assert_eq!(nearest.len(), 1);
-    /// assert_eq!(nearest[0].distance, FXD::from_num(0));
-    /// assert_eq!(nearest[0].item, 100);
-    /// ```
-    #[inline]
-    pub fn nearest_n<F>(&self, query: &[A; K], qty: usize, distance_fn: &F) -> Vec<Neighbour<A, T>>
-    where
-        F: Fn(&[A; K], &[A; K]) -> A,
-    {
-        let mut off = [A::ZERO; K];
-        let mut result: BinaryHeap<Neighbour<A, T>> = BinaryHeap::with_capacity(qty);
+    generate_nearest_n!(
+        (r#"Finds the nearest `qty` elements to `query`, using the specified
+distance metric function.
 
-        unsafe {
-            self.nearest_n_recurse(
-                query,
-                distance_fn,
-                self.root_index,
-                0,
-                &mut result,
-                &mut off,
-                A::ZERO,
-            )
-        }
+# Examples
 
-        result.into_sorted_vec()
-    }
+```rust
+    use fixed::FixedU16;
+    use fixed::types::extra::U0;
+    use kiddo::fixed::kdtree::KdTree;
+    use kiddo::fixed::distance::squared_euclidean;
 
-    unsafe fn nearest_n_recurse<F>(
-        &self,
-        query: &[A; K],
-        distance_fn: &F,
-        curr_node_idx: IDX,
-        split_dim: usize,
-        results: &mut BinaryHeap<Neighbour<A, T>>,
-        off: &mut [A; K],
-        rd: A,
-    ) where
-        F: Fn(&[A; K], &[A; K]) -> A,
-    {
-        if KdTree::<A, T, K, B, IDX>::is_stem_index(curr_node_idx) {
-            let node = &self.stems.get_unchecked(curr_node_idx.az::<usize>());
+    type FXD = FixedU16<U0>;
 
-            let mut rd = rd;
-            let old_off = off[split_dim];
-            let new_off = query[split_dim].dist(node.split_val);
+    let mut tree: KdTree<FXD, u32, 3, 32, u32> = KdTree::new();
 
-            let [closer_node_idx, further_node_idx] =
-                if *query.get_unchecked(split_dim) < node.split_val {
-                    [node.left, node.right]
-                } else {
-                    [node.right, node.left]
-                };
-            let next_split_dim = (split_dim + 1).rem(K);
+    tree.add(&[FXD::from_num(1), FXD::from_num(2), FXD::from_num(5)], 100);
+    tree.add(&[FXD::from_num(2), FXD::from_num(3), FXD::from_num(6)], 101);
 
-            self.nearest_n_recurse(
-                query,
-                distance_fn,
-                closer_node_idx,
-                next_split_dim,
-                results,
-                off,
-                rd,
-            );
+    let nearest: Vec<_> = tree.nearest_n(&[FXD::from_num(1), FXD::from_num(2), FXD::from_num(5)], 1, &squared_euclidean);
 
-            // TODO: switch from dist_fn to a dist trait that can apply to 1D as well as KD
-            //       so that updating rd is not hardcoded to sq euclidean
-            rd = rd.saturating_add(
-                (new_off.saturating_mul(new_off)).saturating_sub(old_off.saturating_mul(old_off)),
-            );
-
-            if Self::dist_belongs_in_heap(rd, results) {
-                off[split_dim] = new_off;
-                self.nearest_n_recurse(
-                    query,
-                    distance_fn,
-                    further_node_idx,
-                    next_split_dim,
-                    results,
-                    off,
-                    rd,
-                );
-                off[split_dim] = old_off;
-            }
-        } else {
-            let leaf_node = self
-                .leaves
-                .get_unchecked((curr_node_idx - IDX::leaf_offset()).az::<usize>());
-
-            leaf_node
-                .content_points
-                .iter()
-                .take(leaf_node.size.az::<usize>())
-                .enumerate()
-                .for_each(|(idx, entry)| {
-                    let distance: A = distance_fn(query, entry);
-                    if Self::dist_belongs_in_heap(distance, results) {
-                        let item = unsafe { *leaf_node.content_items.get_unchecked(idx) };
-                        let element = Neighbour { distance, item };
-                        if results.len() < results.capacity() {
-                            results.push(element)
-                        } else {
-                            let mut top = results.peek_mut().unwrap();
-                            if element.distance < top.distance {
-                                *top = element;
-                            }
-                        }
-                    }
-                });
-        }
-    }
-
-    fn dist_belongs_in_heap(dist: A, heap: &BinaryHeap<Neighbour<A, T>>) -> bool {
-        heap.is_empty() || dist < heap.peek().unwrap().distance || heap.len() < heap.capacity()
-    }
+    assert_eq!(nearest.len(), 1);
+    assert_eq!(nearest[0].distance, FXD::from_num(0));
+    assert_eq!(nearest[0].item, 100);
+```"#)
+    );
 }
 
 #[cfg(test)]
