+++ src/float/query/nearest_n.rs
@@ -1,15 +1,16 @@
 use crate::float::kdtree::{Axis, KdTree};
 use crate::float::neighbour::Neighbour;
-use crate::types::{Content, Index};
+use crate::types::{Content, Index, is_stem_index};
 use az::{Az, Cast};
 use std::collections::BinaryHeap;
 use std::ops::Rem;
 
+use crate::generate_nearest_n;
 
-macro_rules! genreate_nearest_n {
-    ($kdtree:ident, $doctest_build_tree:tt) => {
-    doc_comment! {
-    concat!("Finds the nearest `qty` elements to `query`, using the specified
+macro_rules! generate_float_nearest_n {
+    ($doctest_build_tree:tt) => {
+        generate_nearest_n!((
+            "Finds the nearest `qty` elements to `query`, using the specified
 distance metric function.
 # Examples
 
@@ -17,131 +18,26 @@ distance metric function.
     use kiddo::float::kdtree::KdTree;
     use kiddo::distance::squared_euclidean;
 
-    ",  $doctest_build_tree, "
+    ",
+            $doctest_build_tree,
+            "
 
     let nearest: Vec<_> = tree.nearest_n(&[1.0, 2.0, 5.1], 1, &squared_euclidean);
 
     assert_eq!(nearest.len(), 1);
     assert!((nearest[0].distance - 0.01f64).abs() < f64::EPSILON);
     assert_eq!(nearest[0].item, 100);
-```"),
-    #[inline]
-    pub fn nearest_n<F>(&self, query: &[A; K], qty: usize, distance_fn: &F) -> Vec<Neighbour<A, T>>
-    where
-        F: Fn(&[A; K], &[A; K]) -> A,
-    {
-        let mut off = [A::zero(); K];
-        let mut result: BinaryHeap<Neighbour<A, T>> = BinaryHeap::with_capacity(qty);
-
-        unsafe {
-            self.nearest_n_recurse(
-                query,
-                distance_fn,
-                self.root_index,
-                0,
-                &mut result,
-                &mut off,
-                A::zero(),
-            )
-        }
-
-        result.into_sorted_vec()
-    }
-
-    unsafe fn nearest_n_recurse<F>(
-        &self,
-        query: &[A; K],
-        distance_fn: &F,
-        curr_node_idx: IDX,
-        split_dim: usize,
-        results: &mut BinaryHeap<Neighbour<A, T>>,
-        off: &mut [A; K],
-        rd: A,
-    ) where
-        F: Fn(&[A; K], &[A; K]) -> A,
-    {
-        if $kdtree::<A, T, K, B, IDX>::is_stem_index(curr_node_idx) {
-            let node = &self.stems.get_unchecked(curr_node_idx.az::<usize>());
-
-            let mut rd = rd;
-            let old_off = off[split_dim];
-            let new_off = query[split_dim] - node.split_val;
-
-            let [closer_node_idx, further_node_idx] =
-                if *query.get_unchecked(split_dim) < node.split_val {
-                    [node.left, node.right]
-                } else {
-                    [node.right, node.left]
-                };
-            let next_split_dim = (split_dim + 1).rem(K);
-
-            self.nearest_n_recurse(
-                query,
-                distance_fn,
-                closer_node_idx,
-                next_split_dim,
-                results,
-                off,
-                rd,
-            );
-
-            // TODO: switch from dist_fn to a dist trait that can apply to 1D as well as KD
-            //       so that updating rd is not hardcoded to sq euclidean
-            rd = rd + new_off * new_off - old_off * old_off;
-            if Self::dist_belongs_in_heap(rd, results) {
-                off[split_dim] = new_off;
-                self.nearest_n_recurse(
-                    query,
-                    distance_fn,
-                    further_node_idx,
-                    next_split_dim,
-                    results,
-                    off,
-                    rd,
-                );
-                off[split_dim] = old_off;
-            }
-        } else {
-            let leaf_node = self
-                .leaves
-                .get_unchecked((curr_node_idx - IDX::leaf_offset()).az::<usize>());
-
-            leaf_node
-                .content_points
-                .iter()
-                .take(leaf_node.size.az::<usize>())
-                .enumerate()
-                .for_each(|(idx, entry)| {
-                    let distance: A = distance_fn(query, entry);
-                    if Self::dist_belongs_in_heap(distance, results) {
-                        let item = unsafe { *leaf_node.content_items.get_unchecked(idx) };
-                        let element = Neighbour { distance, item };
-                        if results.len() < results.capacity() {
-                            results.push(element)
-                        } else {
-                            let mut top = results.peek_mut().unwrap();
-                            if element.distance < top.distance {
-                                *top = element;
-                            }
-                        }
-                    }
-                });
-        }
-    }
-
-    fn dist_belongs_in_heap(dist: A, heap: &BinaryHeap<Neighbour<A, T>>) -> bool {
-        heap.is_empty() || dist < heap.peek().unwrap().distance || heap.len() < heap.capacity()
-    }
-}}}
-
+```"
+        ));
+    };
+}
 
 impl<A: Axis, T: Content, const K: usize, const B: usize, IDX: Index<T = IDX>>
     KdTree<A, T, K, B, IDX>
 where
     usize: Cast<IDX>,
 {
-    genreate_nearest_n!(
-        KdTree,
+    generate_float_nearest_n!(
         "let mut tree: KdTree<f64, u32, 3, 32, u32> = KdTree::new();
     tree.add(&[1.0, 2.0, 5.0], 100);
     tree.add(&[2.0, 3.0, 6.0], 101);"
@@ -151,13 +47,17 @@ where
 #[cfg(feature = "rkyv")]
 use crate::float::kdtree::ArchivedKdTree;
 #[cfg(feature = "rkyv")]
-impl<A: Axis + rkyv::Archive<Archived = A>, T: Content + rkyv::Archive<Archived = T>, const K: usize, const B: usize, IDX: Index<T = IDX> + rkyv::Archive<Archived = IDX>>
-ArchivedKdTree<A, T, K, B, IDX>
-    where
-        usize: Cast<IDX>,
+impl<
+        A: Axis + rkyv::Archive<Archived = A>,
+        T: Content + rkyv::Archive<Archived = T>,
+        const K: usize,
+        const B: usize,
+        IDX: Index<T = IDX> + rkyv::Archive<Archived = IDX>,
+    > ArchivedKdTree<A, T, K, B, IDX>
+where
+    usize: Cast<IDX>,
 {
-    genreate_nearest_n!(
-        ArchivedKdTree,
+    generate_float_nearest_n!(
         "use std::fs::File;
     use memmap::MmapOptions;
